+++ src/ast/passes/parser.h
@@ -5,6 +5,7 @@
 #include "ast/passes/config_analyser.h"
 #include "ast/passes/deprecated.h"
 #include "ast/passes/field_analyser.h"
+#include "ast/passes/macro_expansion.h"
 #include "ast/passes/map_sugar.h"
 #include "btf.h"
 #include "clang_parser.h"
@@ -21,6 +22,7 @@ inline std::vector<Pass> AllParsePasses(
   std::vector<Pass> passes;
   passes.emplace_back(CreateParsePass());
   passes.emplace_back(CreateConfigPass());
+  passes.emplace_back(CreateMacroExpansionPass());
   passes.emplace_back(CreateDeprecatedPass());
   passes.emplace_back(CreateParseAttachpointsPass());
   passes.emplace_back(CreateParseBTFPass());
@@ -30,6 +32,7 @@ inline std::vector<Pass> AllParsePasses(
   // The source and syntax is reparsed because it uses the `macros_` which are
   // set during the clang parse to expand identifiers within the lexer.
   passes.emplace_back(CreateParsePass());
+  passes.emplace_back(CreateMacroExpansionPass());
   passes.emplace_back(CreateParseAttachpointsPass());
   passes.emplace_back(CreateMapSugarPass());
   return passes;