+++ src/config.cpp
@@ -253,6 +253,7 @@ const std::map<std::string, AnyParser> CONFIG_KEY_MAP = {
   { "use_blazesym", CONFIG_FIELD_PARSER(use_blazesym) },
   { "show_debug_info", CONFIG_FIELD_PARSER(show_debug_info) },
   { "unstable_import", CONFIG_FIELD_PARSER(unstable_import) },
+  { "unstable_macro", CONFIG_FIELD_PARSER(unstable_macro) },
   { "unstable_map_decl", CONFIG_FIELD_PARSER(unstable_map_decl) },
 };